Fixed not finding texture issue

+++ K-Engine/src/K-Engine/Scene.h
@@ -140,7 +140,7 @@ namespace K
 
 		void Render() 
 		{
-			K::Time::startTime = glfwGetTime();
+			K::Time::startTime = (float)glfwGetTime();
 			K::Time::deltaTimeValue = K::Time::startTime - K::Time::endTime;
 			K::Time::endTime = K::Time::startTime;
 			this->GameLoop();
