+++ K-Engine/src/K-Engine/Time.cpp
@@ -1,11 +1,11 @@
 #include "Time.h"
 namespace K 
 {
-	double Time::startTime = 0.0f;
-	double Time::endTime = 0.0f;
-	double Time::deltaTimeValue = 0.0f;
+	float Time::startTime = 0.0f;
+	float Time::endTime = 0.0f;
+	float Time::deltaTimeValue = 0.0f;
 
-	double Time::deltaTime() 
+	float Time::deltaTime()
 	{
 		return K::Time::deltaTimeValue;
 	}
